* Implemented box structuring element in GlMorphologyFilter * Updated GlGradientVolumeGenerator to generate GL_RGB16F textures to save memory

+++ modules/preprocessing/processors/glgradientvolumegenerator.cpp
@@ -83,7 +83,7 @@ namespace campvis {
             inputUnit.activate();
 
             // create texture for result
-            tgt::Texture* resultTexture = new tgt::Texture(0, tgt::ivec3(size), GL_RGB, GL_RGB32F, GL_FLOAT, tgt::Texture::LINEAR);
+            tgt::Texture* resultTexture = new tgt::Texture(0, tgt::ivec3(size), GL_RGB, GL_RGB16F, GL_FLOAT, tgt::Texture::LINEAR);
             resultTexture->uploadTexture();
 
             // activate shader and bind textures
